refactor profile components
